+++ plugins/transformer/jsonflattener/client/spec/spec.go
@@ -1,0 +1,15 @@
+package spec
+
+type Spec struct {
+	Tables []string `json:"tables"`
+}
+
+func (s *Spec) SetDefaults() {
+	if s.Tables == nil {
+		s.Tables = []string{"*"}
+	}
+}
+
+func (*Spec) Validate() error {
+	return nil
+}