+++ plugins/destination/duckdb/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [5.9.17](https://github.com/cloudquery/cloudquery/compare/plugins-destination-duckdb-v5.9.16...plugins-destination-duckdb-v5.9.17) (2024-10-30)
+
+
+### Bug Fixes
+
+* **deps:** Update module github.com/cloudquery/plugin-sdk/v4 to v4.67.1 ([#19467](https://github.com/cloudquery/cloudquery/issues/19467)) ([7c20418](https://github.com/cloudquery/cloudquery/commit/7c20418bd8fbb5b9a74726c90251fcde9a53b94a))
+
 ## [5.9.16](https://github.com/cloudquery/cloudquery/compare/plugins-destination-duckdb-v5.9.15...plugins-destination-duckdb-v5.9.16) (2024-10-22)